ci: migrate MLA tests to test/registered/mla/
Move 6 MLA test files from test/srt/ to test/registered/mla/ with
CI registry decorators:

- test_mla.py (CUDA: stage-b-test-small-1-gpu, AMD: disabled #13107)
- test_mla_deepseek_v3.py (CUDA: stage-b-test-small-1-gpu, AMD: disabled #12574)
- test_mla_flashinfer.py (CUDA: stage-b-test-small-1-gpu)
- test_mla_fp8.py (CUDA: stage-b-test-small-1-gpu)
- test_mla_int8_deepseek_v3.py (CUDA: stage-b-test-small-1-gpu)
- test_flashmla.py (CUDA: stage-b-test-small-1-gpu)

Total CUDA estimated time: ~1545s (~25.8 min)
