+++ test/registered/mla/test_flashmla.py
@@ -1,3 +1,8 @@
+"""
+Usage:
+python3 test/registered/mla/test_flashmla.py
+"""
+
 import unittest
 from types import SimpleNamespace
 
@@ -5,35 +10,40 @@ import requests
 import torch
 
 from sglang.srt.utils import kill_process_tree
+from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.few_shot_gsm8k import run_eval as run_eval_few_shot_gsm8k
 from sglang.test.test_utils import (
+    DEFAULT_MODEL_NAME_FOR_TEST_MLA,
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
     DEFAULT_URL_FOR_TEST,
     CustomTestCase,
     popen_launch_server,
 )
 
+# FlashMLA attention backend tests with MTP speculative decoding
+register_cuda_ci(est_time=230, suite="stage-b-test-small-1-gpu")
 
-class TestFlashinferMLA(CustomTestCase):
+
+class TestFlashMLAAttnBackend(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls.model = "lmsys/sglang-ci-dsv3-test"
+        cls.model = DEFAULT_MODEL_NAME_FOR_TEST_MLA
         cls.base_url = DEFAULT_URL_FOR_TEST
         other_args = ["--trust-remote-code"]
         if torch.cuda.is_available() and torch.version.cuda:
             other_args.extend(
                 [
-                    "--enable-torch-compile",
                     "--cuda-graph-max-bs",
-                    "4",
+                    "2",
                     "--attention-backend",
-                    "flashinfer",
+                    "flashmla",
                 ]
             )
+        # Use longer timeout for DeepGEMM JIT compilation which can take 10-20 minutes
         cls.process = popen_launch_server(
             cls.model,
             cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
+            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH * 2,
             other_args=other_args,
         )
 
@@ -54,10 +64,10 @@ class TestFlashinferMLA(CustomTestCase):
         metrics = run_eval_few_shot_gsm8k(args)
         print(metrics)
 
-        self.assertGreater(metrics["accuracy"], 0.615)
+        self.assertGreater(metrics["accuracy"], 0.60)
 
 
-class TestFlashinferMLAMTP(CustomTestCase):
+class TestFlashMLAMTP(CustomTestCase):
     @classmethod
     def setUpClass(cls):
         cls.model = "lmsys/sglang-ci-dsv3-test"
@@ -68,25 +78,29 @@ class TestFlashinferMLAMTP(CustomTestCase):
                 [
                     "--cuda-graph-max-bs",
                     "4",
+                    "--disable-radix",
                     "--enable-torch-compile",
                     "--torch-compile-max-bs",
                     "1",
                     "--speculative-algorithm",
                     "EAGLE",
+                    "--speculative-draft-model-path",
+                    "lmsys/sglang-ci-dsv3-test-NextN",
                     "--speculative-num-steps",
-                    "3",
+                    "2",
                     "--speculative-eagle-topk",
                     "1",
                     "--speculative-num-draft-tokens",
-                    "4",
+                    "3",
                     "--attention-backend",
-                    "flashinfer",
+                    "flashmla",
                 ]
             )
+        # Use longer timeout for DeepGEMM JIT compilation which can take 10-20 minutes
         cls.process = popen_launch_server(
             cls.model,
             cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
+            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH * 2,
             other_args=other_args,
         )
 
@@ -117,7 +131,7 @@ class TestFlashinferMLAMTP(CustomTestCase):
             "avg_spec_accept_length"
         ]
         print(f"{avg_spec_accept_length=}")
-        self.assertGreater(avg_spec_accept_length, 2.5)
+        self.assertGreater(avg_spec_accept_length, 2.4)
 
 
 if __name__ == "__main__":
