+++ test/registered/mla/test_mla_deepseek_v3.py
@@ -5,6 +5,7 @@ from types import SimpleNamespace
 import requests
 
 from sglang.srt.utils import is_cuda, is_hip, kill_process_tree
+from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
 from sglang.test.few_shot_gsm8k import run_eval as run_eval_few_shot_gsm8k
 from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
@@ -12,6 +13,14 @@ from sglang.test.test_utils import (
     CustomTestCase,
     is_in_ci,
     popen_launch_server,
+)
+
+# DeepSeek-V3 MLA tests with torch compile, FA3, and MTP speculative decoding
+register_cuda_ci(est_time=442, suite="stage-b-test-small-1-gpu")
+register_amd_ci(
+    est_time=221,
+    suite="stage-a-test-1",
+    disabled="see https://github.com/sgl-project/sglang/issues/12574",
 )
