+++ test/registered/mla/test_mla_fp8.py
@@ -2,20 +2,24 @@ import unittest
 from types import SimpleNamespace
 
 from sglang.srt.utils import kill_process_tree
+from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.run_eval import run_eval
 from sglang.test.test_utils import (
-    DEFAULT_MLA_MODEL_NAME_FOR_TEST,
+    DEFAULT_MLA_FP8_MODEL_NAME_FOR_TEST,
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
     DEFAULT_URL_FOR_TEST,
     CustomTestCase,
     popen_launch_server,
 )
 
+# MLA FP8 KV cache test with MGSM evaluation
+register_cuda_ci(est_time=77, suite="stage-b-test-small-1-gpu")
+
 
 class TestMLA(CustomTestCase):
     @classmethod
     def setUpClass(cls):
-        cls.model = DEFAULT_MLA_MODEL_NAME_FOR_TEST
+        cls.model = DEFAULT_MLA_FP8_MODEL_NAME_FOR_TEST
         cls.base_url = DEFAULT_URL_FOR_TEST
         cls.process = popen_launch_server(
             cls.model,
@@ -23,11 +27,8 @@ class TestMLA(CustomTestCase):
             timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
             other_args=[
                 "--trust-remote-code",
-                "--enable-torch-compile",
-                "--torch-compile-max-bs",
-                "4",
-                "--chunked-prefill-size",
-                "256",
+                "--kv-cache-dtype",
+                "fp8_e5m2",
             ],
         )
 
@@ -45,7 +46,7 @@ class TestMLA(CustomTestCase):
         )
 
         metrics = run_eval(args)
-        self.assertGreater(metrics["score"], 0.8)
+        assert metrics["score"] >= 0.8
 
 
 if __name__ == "__main__":
